Add controllers and fix db migration

+++ src/main/java/com/example/library/controller/BookController.java
@@ -21,7 +21,7 @@ public class BookController {
                            @RequestParam(required = false) String description,
                            @RequestParam(required = false) Integer publishedYear,
                            @RequestParam Set<Long> authorIds,
-                           @RequestParam Set<Long> genreIds) {
+                           @RequestParam Set<Integer> genreIds) {
 
         return bookService.createBook(title, description, publishedYear, authorIds, genreIds);
     }
